Factor out initGUITests
