+++ src/test/CMakeLists.txt
@@ -1,3 +1,4 @@
+
 find_package(Boost COMPONENTS unit_test_framework REQUIRED)
 if(NOT Boost_USE_STATIC_LIBS)
 	ADD_DEFINITIONS(-DBOOST_TEST_DYN_LINK)
