+++ src/test/main.cpp
@@ -35,17 +35,6 @@
 
 namespace bfs = boost::filesystem;
 
-void doInitGameRNG(unsigned defaultValue /*= 1337*/, const char* fileName /*= ""*/, unsigned line /*= 0*/)
-{
-#ifdef RTTR_RAND_TEST
-    RANDOM.Init(rand() + defaultValue);
-#else
-    RANDOM.Init(defaultValue);
-#endif
-    if(fileName && fileName[0])
-        std::cout << "Ingame RNG (" << fileName << "#" << line << ")= " << RANDOM.GetCurrentRandomValue() << std::endl;
-}
-
 struct TestSetup
 {
     TestSetup()
